Created modules for 3 methods
