+++ modules/list_all_books.rb
@@ -1,2 +1,13 @@
 module ListAllBooks
+  def list_all_books
+    puts "\n"
+    if @books_list.empty?
+      puts 'The library is empty.'
+    else
+      @books_list.each_with_index do |value, index|
+        puts "#{index})  Title: #{value.title}, Author: #{value.author}"
+      end
+    end
+    puts "\n"
+  end
 end
