+++ modules/list_all_people.rb
@@ -1,2 +1,13 @@
 module ListAllPeople
+  def list_all_people
+    puts "\n"
+    if @people_list.empty?
+      puts 'The list is empty.'
+    else
+      @people_list.each_with_index do |value, index|
+        puts "#{index})  [#{value.class}] Name: #{value.name}, ID: #{value.id}, Age: #{value.age}"
+      end
+    end
+    puts "\n"
+  end
 end
